Question two on Classes & Methods
This is a Clothing Shop  program and it has item categories by using the *enum* function to list the categories. 
The categories are;
 MENS, 
WOMENS, 
BOYS, 
GIRLS and 
BABIES Clothings.
 The user of the program is able to input the ; 
item name;
item number;
item price;
select category of item;
and also check whether the item selected is a cloth or not.
The program uses switch cases to select category based on the users type of dress or cloth he/she is buying.
Finally, the program uses a user prompt to continue (loop the program) or exit program by pressing "1" to loop and any other to exit program.